Rand: Replacing rand() for portable tests
In order to keep tests simple with respect to signal processing code, a portable
Pseudo Random Number Generator was built.
This will ensure that the sequence of random numbers from a given seed are
identical from platform to platform.
As this is not a cryptographic application, this behavior should not be an
issue.
Basic statistics of random source verified.

For source of magic numbers, see glibc.