OscilGen UI uses OSC transportation
The Oscillator Waveforms and spectra are now transported over OSC.
This commit is a bit of a mixed bag, as after reading more of the OscilGen
source it has become obvious that it is not a RT structure at all and similar to
PADnoteParameters it needs to be migrated to some middle layer structure.
Getting the samples is expensive and so is preparing the oscillator to the point
that even one operation in the RT thread can be enough to force it to miss its
deadline (not to mention it generates quite a few memory allocations when
adaptive harmonics are enabled).

With that said once the storage aspect of moving stuff to the middleware layer
is done, reworking the DSP code should be fairly straightforward.
All operations required to get samples for ADnote are linear transformations of
the data generated by prepare and as far as I can see only one operation depends
upon the chosen frequency (adaptive harmonics).
Thus it should be possible to reorder all other computations and place them at
the prepare() stage, thus leaving some simple math with ~4 Ptype-parameters and
one fft behind to get executed in every get() call.

As per what has been added/done in this commit:

- Multimap of paths to widgets has been created to track UI updates
- Memory leak in OscilGenUI fixed
- Ports added all over the place to get access to the locations of the
  oscillators
- OSC panes/groups introduced as a location specifying shorthand
- OSC widget types introduced in conjunction with the multimap
- Undefined behavior of vumeters fixed (uninitialized values)
- Oscilloscope waveform and spectrum widgets are separated out now as directly
  editing them in fluid is too time consuming
  (some additional refactoring could be done here)
- some parts of OscilGen have been refactored
- Large parts of the UI now either have a osc or osc_i member which contains the
  UI's OSC interface
  If multiple copies of the UI are generated, it should be possible to
  manipulate this along with the base path to direct events wherever you might
  want them