OscilGenUI is free of hazardous pointers
- OscilGenUI no longer has any pointer to master or its own OscilGen instance
  for that matter.
- Leaks are patched
- Verbosity is currently increased drastically in the source due to a few
  logical errors

So, not quite everything is taken care of in OscilGen and a few parameters may
have typos yet to be discovered, though it does appear to be running fairly
smoothly. The memory that is being tossed all over the place is now safely
deallocated and all double free issues are handled.
Some of the code here is a bit more manual than desired, but the automagic
should start kicking in now that rtosc has infected the UI.

Next up wiring some of this stuff to MIDI...