Infection Spread to PADnoteUI
- Fixes limit() argument issues
- Disables PADsynth bank scanning due to startup delay
- Converts most of the unique elements in padsynth's GUI to ports
- Debug output colorified
- Widgets can now know the source of their events

Now things are moving along.
The result is a very snappy UI that dumps out an insane amount of debug
information. OscilGenUI is now more or less a free standing entity.
Give it an OSC interface and a base path and it can run with nothing else.
As per PADsynthUI, some of the standard elements (filters/envelopes) need to be
updated before that thing can run on its own, but once that is done then ADsynth
should be quite easy to finish off.
As per what needs to get refactored -
 * the callbacks are currently a bit inconsistent and possibly hazardous
   based upon oscprompt blowing everything up previously
 * There are quite a few plotting style widgets that have been extracted from
   the UI. Combined these could form a very nice Fl_1D_Plot which would
   eliminate some of the issues that have popped up at each one
 * There should be some way of getting the software to dump out its ENTIRE tree.
   I suspect this is halariously large at this point (16 part 16 kit 3 synths 8
   voice 2 oscil 128 harmonics), but it would be nice to see it.
   Perhaps restricting all of those branch sites to the '0' case would make that
   a sane printout
 * Now that the widgets are getting pulled out, it may make sense to create
   src/UI/widgets/ for organizational purposes