ABS: MIDI controls now enabled
While only one control /volume works via the UI, the rest of the connections
should work if the address is specified correctly. The issue is the basic
handling code in Fl_Osc_Tree does not understand the enumerable port names.

For those reading along on this stream of verbose commits, this adds some
utilities from librtosc's examples that simply store a list of midi addresses
and osc addresses. Once a control is 'learned', then whenever the midi
addressing table sees the event it will be able to construct the corresponding
OSC event to get dispatched immediately.

Now, the next step should be enhancing Fl_Osc_Tree and verifying the above
assertions are actually accurate.