ABS: Fl_Osc_Tree Populated with all entries
Not perfect, as this is currently going to result in some events getting applied
in the wrong thread, but one can now connect to ports lodged deeply in the tree
and see the right widgets wiggle around on the GUI side of thigns.
In this case the tested params were
/part0/kit0/adnote/voice0/oscil/basefuncmodulationpar1
/part0/kit0/adnote/voice0/oscil/basefuncmodulationpar2
/part0/kit0/adnote/voice0/oscil/basefuncmodulationpar3

Judging by the lack of Fl_Oscilloscope updates, I'm going to need to fuss with
when callbacks are spawned or include some of the info in the backend.