Oscify VirKeyboard
- Add noteOn, noteOff, setController handlers to Master port table

This was actually a rather quick and small fix that eliminated about a half
dozen references to the dreaded master->mutex.
Right now MasterUI still assumes that it is the root element, so in future
revisions it will need to start prefixing things correctly, but until then this
approach works fine.