Partial OSC Integration with BankUI
- Add BankView Set of widgets
- Remove some of the need of the bank pointer in BankUI.fl
- Introduce some bug inducing behavior

The bank will realistically only be able to load things at the moment
as safe saving has not been implemented yet.
As the bank has gotten more abstract some of the old hacks to get the
rest of the UI to safely refresh are gone.
This _enhancement_ will result in drawing out some old bugs.
In all likelyhood this will take a few commits to fix up, but it is a
needed step in killing off some more pointers.