MiddleWare: Add Check For Unused UI events
Its easy enough to get the backend to spew out all sorts of unneeded messages so
there should be some alert if they are going to /dev/null.
This is certainly a much weaker case than when an event directed to the backend
misses everything as there are some cases where it is expected that a widget
will not be reached (though they are uncommon).