FilterUI: Add OSC layer
While the formant modification UI is still not quite complete, this seems to
work for the most part and now that this module is out of the way, all of the
major submodules for PADsynth, ADsynth, and SUBsynth seem to be out of the way,
thus stripping them of their pointers should not be too hard.
The main missing feature here is formant/vowel knob moving based on other
controls. Some contemplation has been done on this issue, but nothing has been
finalized yet. As another note, the level of debugging output is getting to the
point that it is difficult to see what needs to be seen. Some sort of
lightweight logging setup should be added. Several of the widget classes have
been partially converted to the new style interface, but that will still take a
while to complete.