Partial Oscification of PartUI/MasterUI
This is a bit of a half baked commit as significant work needs to be done on the
side of MiddleWare and Effects before either of these modules can be finished
off. As such there are a number of dangling widgets which may or may not get
updated to the right location on a rebase() operation.

Notably:

- MiddleWare support of deleting parts is needed to continue
- Microtonal/Controls needs to be validated
- Fl_Osc_Input needs to be properly implemented
- Some custom widget is needed for the bank UI spawner
- the basic UI needs to be tested *shudder*