SUBnote: 2x Speed Enhancement
Well, SUBnote is now a LOT faster. According to the SUBnoteTest
the time to compute samples is about halved now due to a change
in the filtering code.
Time for 15,000 samples went from 2.75 sec to 1.34 sec under
the BuildOptionsBasic configuration.
This should reflect real usage as both cases were compute bound
in the filtering loop.