Enable Lockless Saving XMZ Files
While this code isn't going to make any thread error detection tool too happy,
there are now memory barriers in place to freeze the backend and propagate any
pending changes to whatever coherency model is needed to ensure that the
middleware is able to grab the parameters safely from memory without a copy.
This general approach shall be extended to the cases of the parameters
responsible for Part, ADnote, SUBnote, PADnote, and perhaps ADnoteVoice.

The oddest part of this commit would have to be the fire in order complete
(slightly) out of order buffer needed to ensure that no events were lost on a
frozen state, but now that it seems to be sorted, doing this operation for other
structures should be very easy.

- Add code needed to transition backend into parameter readonly state
- Add handler for saving xmz files