Enable External UI Support
This adds support for an external UI with zero access to the core access.
All access is done through OSC messages which can be either on a local or remote
machine, though there may be some minor cases where they depend on struct
packing due to shoving stuff into blobs.
There doesn't seem to be any lag given that both sides consume all events when
they check for them and when running with a large amount of latency elements
still appear to be correct in the end.

This does add some regressions in the controls which did depend upon direct
access, such as the ConfigUI, PresetUI, and NioUI, however those should be fixed
to use the OSC API anyhow. Some new widget types will be made for these cases
and the vision of everything over OSC will be obtained.