UI: Capture Messages During Initialization
Previously during the construction of the GUI a number of messages
would receive responses, however these would not be routed to widgets
as the gui callback was not set until after initialization.
These messages are now captured in a temporary vector and played back
to the UI after setting the callback.
This should fix some startup weirdness which has been visible
in the past.