Fix for use after free in XMLwrapper.
If loading of a new XML file fails, the node and root pointers
can still be set pointing to freed elements. This happened
sometimes in the lv2-ttl-generator during build under FreeBSD.

The reason is that some current code tries to enterbranch() even
if loadXMLfile() has failed, and then the a stale node
pointer is followed. For example in BankDb.cpp:

> 131	    xml.loadXMLfile(getCacheName());
> 132	    if(xml.enterbranch("bank-cache")) {

Signed-off-by: Hans Petter Selasky <hps@selasky.org>