Add Offline Mode To Middleware For Plugin Version
Middleware assumes that the realtime thread is consistently running.
If a plugin host 'intelligently' decides to disable running the audio
thread then middleware previously assumed the thread must have crashed.
As a workaround if the audio thread doesn't respond to timing pulses fast
enough, then it's assumed to be offline and Middleware pumps the OSC queue
manually and it performs read only operations without waiting.

This commit is the first pass at implementing this feature and some bugs
are to be expected.