src/CMakeLists.txt: use the target compiler during config tests
The availability of the "fistpl" x86 instruction was tested with a
direct call to the "as" command. In case the project is cross-compiled
to another architecture, the host compiler is tested instead the
target one. This was leading to compilation failures. This commit now
perform the test using the target compiler defined in CMake, which
fixes cross-compilation. This commit does not change the behavior
when doing a native compilation, as the host and target compiler are the
same.

Signed-off-by: Julien Olivain <ju.o@free.fr>