Change Travis-CI Build To Use Distro Liblo Version
Ubuntu Xenial has version 0.28 of liblo, so we shouldn't need to
build it from source. This should resolve a SSL connection error
during the build. It's unclear if that error in particular is a
result of sourceforge having an underlying issue or a problem with
the SSL library available on travis's side.