Fix crash when loading while backend is dead
MiddleWare can request Master to reload (from a blob), and it can also take
over OSC message handling of the Master if the backend is dead. If the
former happens while the backend is dead, and then the latter happens,
MiddleWare by now calls Master::runOsc() on the new master, telling it to
exchange with the new master, and then deleting the new master.
Consequently, MiddleWare remains with a dangling Master pointer, leading
to segfaults in DAWs where the audio thread is not active while loading the
plugin preset.

If now MiddleWare runs Master, it passes its previous Master to runOsc,
so runOsc then considers the previous master as "this", so it does a
real exchange and requests to free the previous master.