Implement workaround for old presets.
All effect parameters have a default value.
Default values are skipped when storing parameters,
and must appear as the default value when loading.

Up until recently it was assumed that the default
value of all parameters is zero. This is no longer
true, but when loading old presets we need to
preserve this behaviour! Else sounds may change.

Add exception for the recently added DC offset parameter.

Bump the ZynAddSubFX version.

Signed-off-by: Hans Petter Selasky <hps@selasky.org>