Remove Controller::operator=
`Controller` contains a member reference, so the default constructor
would not work. The previous implementation used `memcpy` as a
workaround for that in `Controller::operator=`. This triggered a
compiler warning that `memcpy` of an object with non-trivial
copy-assignment should be avoided.

This commit removes this `Controller::operator=` workaround, and clones
the `Controller` object by destroying it explicitly and then using the
copy ctor via placement new.

Furthermore, remove `Controller::~Controller` (Rule of zero).