Change which table AD fills (non-functional)
Implements 4aeee67 "Communication: Change which table AD fills"
from github.com/JohannesLorenz/zyn-wavetable-concept.

In "set-waves", the criterion "semantic index not passed" is currently
equivalent to "is parameter change". We split this up, for future
commits: We will now use "semantic index not passed" to decide which
wave to take, and use "is parameter change" to check the parameter
change time stamps.