Move prepare:b to where it has been called
Prior to this commit, OscilGen's prepare/magnitude/phase ports prepare
data and send it to prepare:b. Since we have no RT thread in OscilGen
anymore, this commit does not send anything to the RT side anymore -
instead, code where the new FFT data has been sent is replaced by the
content of the prepare:b port.

This fixes memory leaks, because prepare:b did not "/free" the freqs
prior to this commit.