Update DPF
This commit should update from a 2016 version of DPF, so for plugins
this should hopefully improve plugin host compatiability.
While I've confirmed that this commit builds on Linux, I haven't tested
the execution, so some regressions may have occured, but the general
expectation is that this will be a step forward.