Implement per-note portamento (#304)
Implement per-note portamento, rather than per-controller (in reality,
per-part).

For each note played, when portamento is needed, create a portamento
object which is referred to by all constituent synth engines, and which
is updated by the part continually. The portamento object is deleted
by the note pool when a note is killed, by which time the synth engines
will have been deallocated and there exists no reference to it.

For legato mode, reuse the portamento object, since the note
descriptors are also reused. Otherwise, the portamento object would
have to be deallocated in various places in the note pool code whenever
a note was reused in legato mode, which would clutter up the code,
and possibly lead to a number of corner cases where the portamento
object was still referred to by the synth engines.
On the other hand, the part code would become cleaner if the
portamento object were not reused.

Whenever a new note starts its glide when portamento is enabled,
start at the current pitch of the previously played note
(including any portamento the previous note is currently
exhibiting), rather than simply the target pitch of the previous
note.

This will create a much smoother portamento, especially in legato
mode, with no pitch jumping, which was the case previosly when
new notes started their portamento cycle at the target pitch of
the previous note.

If the previous note is killed before its portamento cycle has
completed, save the pitch of the note when it was killed
(or, more exactly, when the portamento object is destroyed, which
should be virtually the same time) to use as the starting pitch
for the portamento cycle of the next note played. This is a slightly
rare corner case, but fairly simple to handle.

Portamento is now tested in its own test PortamentoTest,
which means ControllerTest is now empty. Leave it as it is for
future Controller test initiatives.