Fix part mute in jack multi mode (#327)
When part was muted, it was still sent on the individual output channels
when operating in jack-multi mode, which resulted in the last audio
buffer that had been generated beeing looped.

This turned out to be slightly more complex to fix than expected.
First of all, when a part is muted, Part::AllNotesOff() is called,
which makes the next call to Part::ComputePartSmps() gracefully shut
down the part, clearing all notes, resetting the part insert effects,
and ramping down the signal to zero in the buffer being generated to
avoid clicks. However, as the code was written, ComputePartSmps() was
never called for a disabled part, so the graceful shutdown was not
executed until the part was subsequently enabled. So simply setting the
output buffer from the part to zero when the part is disabled is not
sufficient, as there will be a slight click when it is subsequently
re-enabled, and ComputePartSmps() finally gets a chance to go through
its shutdown sequence, after having essentially been in suspended
animation while the voice was disabled.

The solution is to always call ComputePartSmps(), which will then
internally decide when to execute the audio generation code or whether
to generate silence.

When generating the master audio bus, there are still a lot of places
where the audio is chopped off when Penabled for the part is set to
false, resulting in a more audible click than technically should be
necessary.  This could be addressed in a separate patch if needed.