Bank.cpp: Update database after write (#331)
When writing a patch to a bank slot, rescan the database to keep
it updated. Otherwise, a subsequent search() call will not reflect
the current state of the banks, and consequently switching to Read
mode in the UI after writing to a slot in Write mode causes the old
bank contents to be displayed, and also causes any attempt to
load the recently saved patch to fail, as the corresponding patch
file no longer exists.