MiddleWare.cpp: Send OSC type tag ,N for empty bank search
With the following commit in Zest:
"osc-bridge: Only treat 2+ arg messages as vectors"
/bank/search_results messages with no arguments, as the result
of the search yielding no results, will be ignored by osc-bridge,
and consequently the patch list will not be updated when the
search results is an empty list, ultimately resulting in the patch
column erroneously displaying whatever was the result of the
previous search.

Fix this by emitting a ,N (Nil) OSC type tag indicating that
there is actually no argument data, when the search result
is empty. This message will be correctly interpreted by osc-bridge
in Zest, resulting in a zero length write to the cache, and
subsequent display of the search results in the form of an
empty preset column, as intended.