Update controller values when amounts changed
Similar as was recently done for the modwheel, update the corresponding
parameter values not just when the MIDI CC controlling the parameter
is received, but also when the corresponding amount is changed.
Otherwise the user experience will be confusing, as a new controller
value must be received before a depth change can be noticed.