Update controller values when enable switches changed
Similar to updating controller values when depths are changed,
update those that (only) have enable switches when the corresponding
switch is changed. Otherwise, when an enable switch is changed,
the underlying parameter value will remain at its current
value until a new MIDI CC value is received. Especially
in the case of disabling external control, this is counter
intuitive as it means that the underlying parameter does
not change back to its default value until a new MIDI CC
(whose value is now irrelevant) is received.