AnalogFilter: Let setfreq_and_q() update coeeffs when needed (#347)
For the Analog filter (only), only the frequency was updated when
the setfreq_and_q() method was called. The consequence of this was
that updating the Q would only have an effect if for some reason
the frequency was being updated at the same time, for instance
because it was being swept by the envelope.

This also fixes the problem that changing MIDI CC#71 (filter Q)
would not have any effect until MIDI CC#74 (filter freq) was also
changed.

Fix by recomputing the coefficients if Q was changed, and just
updating the frequency if it was not.