Implement own thread for rendering samples
This functionality is controlled by the HAVE_BG_SYNTH_THREAD cmake keyword.
The default value is off.

Some audio backend drivers like JACK may have insanely small buffers which
must be returned to the audio driver in a fashion that leaves no time for
ZynAddSubFX to produce the required audio output, leading to audio underruns,
which does not sound good. Unfortunatly JACK does not support per-client
buffering.

Add generic functionality to the OutMgr class to render audio in the
background, returning already generated audio as quick as possible.

While at it add an enum, FRAME_SIZE_MAX, which basically sets a hard
limit on the allowed per-tick frame size values support. Assert this,
because violations means writing outside of the designated buffers.
This value has been raised from 4K to 16K, which also reduces the amount
of copying when the staled samples needs to be cleaned up.

Make a slight change to the MIDI input algorithm, to reset the time window
it is looking at, when the end of MIDI events is seen. This means MIDI events
received faster than synth.buffersize samples, will be considered received
at the same time.

Fix #216

Signed-off-by: Hans Petter Selasky <hps@selasky.org>