Fixed compilation of zynaddsubfx-ext-gui on MacOS
Original issue: `make zynaddsubfx-ext-gui` failed because it couldn't find the png library. The -lpng flag is generated by fltk-config, which apparently asssumes that it's installed system-wide and doesn't require a -L flag. This assumption is incorrect when libpng has been installed through Homebrew.

The fix here is to pass the homebrew lib/ directory to the linker, when homebrew is installed.