Implement/use transmitMsgGui
This fixes an issue only related to in-process remotes, aka "GUI".

`MwDataObj::reply` uses `MiddleWareImpl::sendToCurrentRemote`, which
sometimes uses `MiddleWareImpl::last_url` as the reply URL. However,
the latter variable is only being set in `handler_function`, i.e. in
the liblo callback, so in case the GUI was last, `last_url` has not
been updated, and `MwDataObj::reply` sends to "nowhere".

In the other case, `MiddleWareImpl::sendToCurrentRemote` makes use
of `MiddleWareImpl::current_url` as the reply URL. This variable is
only being set by an `echo` reply from Master, so `echo` needs to
be sent to `MiddleWare`, which - once again - is only being done in
`handler_function` before this patch.

This commit provides `transmitMsgGui*` functions, which are like the
`transmitMsg*` functions, except that they set `last_url` and
`current_url` accordingly. All usages of `transmitMsg*` shall be
transferred to `transmitMsgGui` if being done in GUI context and if
there is a chance that you care about the answer.

* UI/:  Wherever we call `MiddleWare::transmitMsg` (not the case in
        Connection.cpp)
* Tests/: Only in SaveOSC.cpp (because the answer is relevant
* PresetExtractor.cpp: Nowhere, because all calls are in callbacks
        which were already preceeded by a call to either
        `transmitMsg*` or `transmitMsgGui*`, which has already set
        `last_url` correctly
* MiddleWare: Nowhere, because all calls are either in the
        `handler_function`, or they are in callbacks like
        `loadPart`, which were already preceeded by a call to
        either `transmitMsg*` or `transmitMsgGui*` which has
        already set `last_url` correctly
* main.cpp: Nowhere, because a reply does not make sense here