Allow 2 in-process UIs
This allow MiddleWare to exchange OSC messages with a further, second
in-process UI. This will be needed for a further commit where
both PortChecker's clients will be in-place UIs.

Since this is only needed for PortChecker, the second UI is reduced to
OSC messaging and on purpose has no `Fl_Osc_Interface`.